project/settings.py
import os
from dotenv import load_dotenv
from environs import Env

load_dotenv(override=True)
env = Env()

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql_psycopg2',
        'HOST': os.getenv("HOST"), #'checkpoint.devman.org',
        'PORT': os.getenv("PORT"), #'5434',
        'NAME': os.getenv("NAME"), #'checkpoint',
        'USER': os.getenv("USER"), #'guard',
        'PASSWORD': os.getenv("PASSWORD"), #'osim5',
    }
}

INSTALLED_APPS = ['datacenter']

SECRET_KEY = 'REPLACE_ME'

# DEBUG = False
DEBUG = env.bool("DEBUG")

ROOT_URLCONF = 'project.urls'

ALLOWED_HOSTS = ['*']


BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [os.path.join(BASE_DIR, 'templates')],
        'APP_DIRS': True,
    },
]


USE_L10N = True

LANGUAGE_CODE = 'ru-ru'

TIME_ZONE = 'Europe/Moscow'

USE_TZ = True

DEFAULT_AUTO_FIELD = 'django.db.models.AutoField'
